pyproject.toml
[tool.poetry]
name = "books_library_gateway"
version = "0.0.0"
description = "Service to work with books library"
authors = ["Nosov Pavel <paul-36@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
tomli = "^2.0.1"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
invoke = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
